requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
WTForms>=3.1.0
email-validator>=2.0.0
Pillow>=10.0.0
Werkzeug>=3.0.0

# 图像识别（迁移学习训练与推理）；CPU 安装: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
